requirements.txt
boto==2.49.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
distro==1.8.0
google-api-core==2.11.1
google-auth==2.21.0
google-cloud==0.34.0
google-cloud-compute==1.11.0
google-cloud-core==2.3.2
google-cloud-storage==2.10.0
google-compute-engine==2.8.13
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-status==1.56.0
idna==3.4
proto-plus==1.22.3
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
requests==2.31.0
rsa==4.9
six==1.16.0
urllib3==1.26.16
